+++ src/css/Homepage.css
@@ -1,3 +1,9 @@
+*{
+  padding:0;
+  margin:0;
+  box-sizing: border-box;
+}
+
 .btn_signup{
   border-radius: 40px;
 }
